* src/mcs51/gen.c (genFunction, genEndFunction), * src/ds390/gen.c (genFunction, genEndFunction): Moved save and restore of EA so that stack offsets to parameters are correct when using both critical and reentrant/stack-auto. * src/z80/gen.c (aopOp): removed erroneous assertion about sloc size (can be triggered in error if sloc is shared between different sized objects) * device/include/float.h: fixed macros to explicitly use unsigned long where needed
git-svn-id: https://sdcc.svn.sourceforge.net/svnroot/sdcc/trunk/sdcc@2892 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ device/include/float.h
@@ -41,12 +41,12 @@
 /* the following deal with IEEE single-precision numbers */
 #define EXCESS		126
 #define SIGNBIT		((unsigned long)0x80000000)
-#define HIDDEN		(unsigned long)(1 << 23)
-#define SIGN(fp)	((fp >> (8*sizeof(fp)-1)) & 1)
-#define EXP(fp)		(((fp) >> 23) & (unsigned int) 0x00FF)
+#define HIDDEN		(unsigned long)(1ul << 23)
+#define SIGN(fp)	(((unsigned long)(fp) >> (8*sizeof(fp)-1)) & 1)
+#define EXP(fp)		(((unsigned long)(fp) >> 23) & (unsigned int) 0x00FF)
 #define MANT(fp)	(((fp) & (unsigned long)0x007FFFFF) | HIDDEN)
 #define NORM            0xff000000
-#define PACK(s,e,m)	((s) | ((e) << 23) | (m))
+#define PACK(s,e,m)	((s) | ((unsigned long)(e) << 23) | (m))
 
 float __uchar2fs (unsigned char);
 float __schar2fs (signed char);
